requirements.txt
Pillow
cryptography
tkinterdnd2